feat: create modal window for add & edit
