+++ src/css/modal.css
@@ -1,94 +1,93 @@
 .modal-layer {
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: rgba(0, 0, 0, 0.5);
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 100%;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: rgba(0, 0, 0, 0.5);
 }
 
 .modal-content {
-    position: relative;
-    padding: 25px;
-    background: white;
-    border-radius: 16px;
-    height: 60%;
-    min-height: 500px;
-    width: 40%;
-    margin: auto;
-    display: flex;
-    flex-direction: column;
+  position: relative;
+  padding: 25px;
+  background: white;
+  border-radius: 16px;
+  height: 60%;
+  min-height: 500px;
+  width: 40%;
+  margin: auto;
+  display: flex;
+  flex-direction: column;
 }
 
 .item-note__header {
-    font-size: 2em;
-    font-weight: 600;
-    text-align: center;
-    margin: 0 0 20px 0;
+  font-size: 2em;
+  font-weight: 600;
+  text-align: center;
+  margin: 0 0 20px 0;
 }
 
 .item-note__body {
-
-    margin: 0 0 20px 0;
+  margin: 0 0 20px 0;
 }
 
 .item-note__lines {
-    margin: 0 0 10px 0;
+  margin: 0 0 10px 0;
 }
 
 .item-note__footer {
-    display: flex;
-    justify-content: center;
+  display: flex;
+  justify-content: center;
 }
 
 .item-note-btn__footer {
-    width: 20%;
+  width: 20%;
 }
 
-.item-note_btn {
-    padding: 10px 15px;
-    font-size: 1.1rem;
-    border: 1px solid black;
-    border-radius: 5px;
-    width: 100px
+.item-note__btn {
+  padding: 10px 15px;
+  font-size: 1.1rem;
+  border: 1px solid black;
+  border-radius: 5px;
+  width: 100px;
 }
 
 select {
-    border: 1px solid teal;
-    border-radius: 2px;
-    font-size: 1.2rem;
-    width: 100%;
-    padding: 5px 15px;
-    margin: 5px 0;
+  border: 1px solid teal;
+  border-radius: 2px;
+  font-size: 1.2rem;
+  width: 100%;
+  padding: 5px 15px;
+  margin: 5px 0;
 }
 
 option[selected] {
-    color: white;
-    background-color: teal;
+  color: white;
+  background-color: teal;
 }
 
 .item-note__text {
-    font-size: 15px;
-    font-weight: 300;
-    text-align: left;
-    color: teal;
+  font-size: 15px;
+  font-weight: 300;
+  text-align: left;
+  color: teal;
 }
 
 .item-note-value__text {
-    font-size: 16px;
-    font-weight: 500;
-    text-align: left;
-    color: teal;
-    padding: 5px 15px;
-    margin: 5px 0;
+  font-size: 16px;
+  font-weight: 500;
+  text-align: left;
+  color: teal;
+  padding: 5px 15px;
+  margin: 5px 0;
 }
 
 .item-note__input {
-    width: 100%;
-    padding: 5px 15px;
-    margin: 5px 0;
-    border: 1px solid teal;
+  width: 100%;
+  padding: 5px 15px;
+  margin: 5px 0;
+  border: 1px solid teal;
 }
