--- src/modules/components/createNote.js
@@ -1,1 +1,0 @@
-export function createNote() {}
