+++ src/modules/components/list/createNote.js
@@ -1,0 +1,7 @@
+import { showModalNote } from '../modal/showModalNote';
+import { Note } from '../../models/note';
+
+export function createNote() {
+  const note = new Note();
+  showModalNote(note);
+}
