+++ src/modules/components/list/createWrapper.js
@@ -1,0 +1,31 @@
+import { createElement } from '../../helpers/createElement';
+import noteService from '../../services/noteService';
+import { btnCreateNote } from './createBtnCreate';
+import { createCategoryHeader } from './createCategoryHeader';
+import { createHeader } from './createHeader';
+import { renderListNotes } from './renderListNotes';
+import { renderCategoryNotes } from './renderCategoryNotes';
+
+export function createWrapper() {
+  const rootWrapper = createElement({ tagName: 'div', className: 'wrapper' });
+  const listWrapper = createElement({
+    tagName: 'div',
+    className: 'list-notes__wrapper',
+  });
+
+  listWrapper.appendChild(createHeader());
+  listWrapper.appendChild(renderListNotes(noteService.getNotes()));
+  listWrapper.appendChild(btnCreateNote());
+
+  const categoriesWrapper = createElement({
+    tagName: 'div',
+    className: 'categories-notes__wrapper',
+  });
+  categoriesWrapper.append(createCategoryHeader());
+  categoriesWrapper.append(renderCategoryNotes(noteService.getStat()));
+
+  rootWrapper.appendChild(listWrapper);
+  rootWrapper.appendChild(categoriesWrapper);
+
+  return rootWrapper;
+}
