+++ src/modules/components/modal/showModalNote.js
@@ -3,15 +3,13 @@ import { updateModal } from './updateModal';
 import '../../../css/modal.css';
 
 export function showModalNote(note) {
-    let modal = document.querySelector('.modal-layer');
-    if (!modal) {
-        const root = document.getElementById('root');
-        modal = createModal(note);
-        root.append(modal);
-    } else {
-        
-    }
+  let modal = document.querySelector('.modal-layer');
+  if (!modal) {
+    const root = document.getElementById('root');
+    modal = createModal(note);
+    root.append(modal);
+  }
 
-    modal.style.display = 'flex';
-    updateModal(note);
+  updateModal(note);
+  modal.style.display = 'flex';
 }
