+++ src/modules/models/note.js
@@ -1,10 +1,10 @@
 export class Note {
   constructor(id, name, category, content) {
     this.id = id;
-    this.name = name;
+    this.name = name ? name : '';
     this.created = new Date();
     this.category = category;
-    this.content = content;
+    this.content = content ? content : '';
     this.archived = 0;
   }
